move equipment to module subfolder
