+++ controller/modules/ato/controlller.go
@@ -8,7 +8,7 @@ import (
 	"github.com/reef-pi/types"
 
 	"github.com/reef-pi/reef-pi/controller/connectors"
-	"github.com/reef-pi/reef-pi/controller/equipment"
+	"github.com/reef-pi/reef-pi/controller/modules/equipment"
 	"github.com/reef-pi/reef-pi/controller/utils"
 )
