+++ controller/modules/equipment/controller_test.go
