+++ controller/modules/equipment/equipment.go
