+++ controller/modules/timer/controller.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/reef-pi/types"
 
-	"github.com/reef-pi/reef-pi/controller/equipment"
+	"github.com/reef-pi/reef-pi/controller/modules/equipment"
 )
 
 type Controller struct {
